Cart compatability checks upgrade.
Added a function so that when you remove an item from your cart, it also updates the payment type session, formerlly named checkout_payment_type().

Renamed the function to _checkCartCompatibility().

Made the function private so that it doesn't clutter up the acos table needlessly.

These are all good updates because it uses the naming conventions, and keeps the db lighter.